feat: auth protection for articles endpoint
protected:
- createArticle, upgradeArticle, deleteArticle

fix: authenticationMiddleware typo

+++ controllers/articleController.js
@@ -1,19 +1,15 @@
 import BadRequestError from '../errors/badRequestError.js';
+import UnauthorizedError from '../errors/unauthorizedError.js';
 import { ArticleModel } from '../models/articles.js';
 
 export const createArticle = async (req, res) => {
-  const { text, title, authorId } = req.body;
+  const { text, title } = req.body;
+  const { _id: authorId } = req.user;
 
-  if (!text || !authorId || !title) {
+  if (!text || !title) {
     throw new BadRequestError(
       `${
-        !text
-          ? 'Text is missing'
-          : !title
-          ? 'Title is missing'
-          : !authorId
-          ? 'Author id is missing'
-          : 'Bad request'
+        !text ? 'Text is missing' : !title ? 'Title is missing' : 'Bad request'
       }`
     );
   }
@@ -21,6 +17,7 @@ export const createArticle = async (req, res) => {
   try {
     const article = await ArticleModel.create({
       text,
+      title,
       authorId,
     });
 
@@ -53,7 +50,14 @@ export const upgradeArticle = async (req, res) => {
   const id = req.params.id;
   if (!id) throw new BadRequestError('Article Id not provided');
 
+  const { _id: authorId } = req.user;
+
   try {
+    const article = await ArticleModel.findById(id);
+
+    if (article?.authorId !== authorId)
+      throw new UnauthorizedError('Not authorized to access this route');
+
     await ArticleModel.findByIdAndUpdate(id, { text });
 
     res.status(200).send();
@@ -66,7 +70,14 @@ export const deleteArticle = async (req, res) => {
   const id = req.params.id;
   if (!id) throw new BadRequestError('Article Id not provided');
 
+  const { _id: authorId } = req.user;
+
   try {
+    const article = await ArticleModel.findById(id);
+
+    if (article?.authorId !== authorId)
+      throw new UnauthorizedError('Not authorized to access this route');
+
     await ArticleModel.findByIdAndDelete(id);
 
     res.status(204).send();
